Correct mistake calculating the minimum positive value of the input data.
git-svn-id: https://svn.code.sf.net/p/pymca/code@1362 a8702ac3-0063-49a2-b157-4a214673c9c1

+++ PyMca/sps/Src/spslut_py.c
@@ -1,5 +1,5 @@
 #/*##########################################################################
-# Copyright (C) 2004-2010 European Synchrotron Radiation Facility
+# Copyright (C) 2004-2011 European Synchrotron Radiation Facility
 #
 # This file is part of the PyMCA X-ray Fluorescence Toolkit developed at
 # the ESRF by the Beamline Instrumentation Software Support (BLISS) group.
@@ -202,8 +202,8 @@ static PyObject *spslut_transform(PyObject *self, PyObject *args)
   }
 
   data = src->data;
-  cols = src->dimensions[1];  /*FIX THIS cols and rows are turned around */
-  rows = src->dimensions[0];  /*###CHANGED - ALEXANDRE 24/07/2001*/
+  cols = (int) src->dimensions[1];  /*FIX THIS cols and rows are turned around */
+  rows = (int) src->dimensions[0];  /*###CHANGED - ALEXANDRE 24/07/2001*/
 
 
 
@@ -337,8 +337,8 @@ static PyObject *spslut_transformarray(PyObject *self, PyObject *args)
   }
 
   data = src->data;
-  cols = src->dimensions[1];  /*FIX THIS cols and rows are turned around */
-  rows = src->dimensions[0];  /*###CHANGED - ALEXANDRE 24/07/2001*/
+  cols = (int) src->dimensions[1];  /*FIX THIS cols and rows are turned around */
+  rows = (int) src->dimensions[0];  /*###CHANGED - ALEXANDRE 24/07/2001*/
 
   r = SPS_PaletteArray (data, type, cols, rows, reduc, fastreduc, meth, gamma,
             autoscale, mapmin, mapmax, Xservinfo, palette_code,
